+++ src/test/resources/com/pestphp/pest/Pest.php
@@ -1,0 +1,6 @@
+<?php
+
+namespace Tests;
+
+uses(TestCase::class)->in('Feature');
+uses(TestCase::class)->in('Unit');