+++ src/test/resources/com/pestphp/pest/generators/ExpectationGenerator/GeneratedWithMethod.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Pest\Expectations;
+
+/**
+ * @method bool works()
+ */
+final class Expectation {}