+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ## [Unreleased]
 ### Added
+- Added structure support for tests
+- Added autocompletion for custom expectations
+- Added pest icon for the Pest.php config file
+
+### Fixed
+- Fixed a read only permission bug when used with Code with me
+
+## [0.4.3]
+### Added
 - Added IntelliJ version to bug report
 - Added new Dataset icons (Thanks @caneco!)
 - Added test state icons
@@ -14,7 +23,7 @@
 - Fix running tests with circumflex (^)
 
 ### Changed
-- Bumped min IntelliJ version t0 2020.3
+- Bumped min IntelliJ version to 2020.3
 
 ## [0.4.2]
 ### Added